Exit with non-zero exit code on error
A change in exit code behavior was introduced in #166, where printing an
error used to also cause the application to exit with a non-zero exit
code. As a result, errors in e.g. `kpt live apply` exits the process
with a 0 exit code.

This change should bring it in line with how it works before, but I'm open to
other solutions if there's a cleaner way to accomplish this without
having an error log automatically mean that the application needs to
exit.