Allow colons in object names
In most objects, kubernetes does not allow colons, but especially in
Roles and Bindings they are widely used. On the other hand, colons are
not allowed as keys of ConfigMaps. When we try to write e.g. a
ClusterRole named `psp:rook`, `ObjectMetadata.String` results in a name
like `_psp:rook_rbac.authorization.k8s.io_ClusterRole`, which
consequently is refused by the apiserver.

This commit translates the colon into a underscore, so that the result
of `ObjectMetadata.String` is
`_psp_rook_rbac.authorization.k8s.io_ClusterRole`.