Add option for custom status readers
This commit adds the `CustomStatusReadersFactoryFunc` option to the
StatusPoller's `Poll` method options so that consumers can provide
custom status readers for certain GroupKinds (esp. CRDs).

It also changes the `NewGenericStatusReader` func so that it accepts a
statusFunc parameter so that consumers only have to implement that
single func and use the genericStatusReader for all the boilerplate
stuff (like fetching an object from the API).

refs #382