Allow the users to specify whether to continue on error
Today, cli-utils continues on all errors with one exception: it
terminates immediately on a WaitTask TimeoutError.

This PR allows the users to specify whether to continue on error through
a bool-type field named `ContinueOnError` of the applier/destroyer
Options. The default value is false: cli-utils would terminate
immediately on a WaitTask TimeoutError. If the field is set to true,
cli-utils would not terminate, but instead log the error and send the
error to the event channel before continuing.