feat: rewrite inventory client with basic CRUD
This change rewrites the inventory client interface using simple CRUD
methods. Prior to this change the inventory client consisted of several
tightly coupled interfaces which made the client more complicated and
less flexible.

The old client also performed a Get of the inventory many times
throughout the applier/destroyer processes, including before every
update. This approach discarded optimistic concurrency protection since
the resourceVersion was never persisted. The new approach preserves an
internal interface representation of the inventory so that the
resourceVersion can be preserved across updates.

The new client interface is simpler to implement and is intended to
support implementations for sharded inventories, which may consist of
multiple underlying KRM objects. The prior interfaces made assumptions
of a single KRM object in several places, which made it incompatible
with a sharded inventory.