refactor: return error from ConfigMap wrap functions
This change is primarily motivated by a subsequent change which replaces
the lazy conversion with an immediate conversion, which requires for an
error to be returned. Since this touches so many pieces of code this is
broken out as a separate change.

This also fixes several places where errors were not being propagated.