refactor: optimize UnstructuredClient performance
The UnstructuredClient currently uses the same ToUnstructuredFunc to
generate the spec and the status. This is inefficient when status is
a subresource, because the ToUnstructuredFunc implementation always
needs to return both the spec and the status, which means they both
get sent over the wire to the Kubernetes API. The Kubernetes API will
ignore the spec when using the status endpoint, and ignore the status
when using the normal resource endpoint, so doing all the extra work
to generate, copy, and format the extra fields is inefficient on the
client side. In some edge cases this can also cause errors on the
server side, when the spec and status are larger than the maximum
API request size limit but smaller than the etcd object storage size
limit.

To work around this issue, a new toStatus ToUnstructuredFunc can be
injected into the UnstructuredClient. When non-nil, this new func will
be used before calling UpdateStatus.

This has the added benefit that it makes it possible to use the
UnstructuredClient on custom inventory objects that have a status field
but not a status subresource.

To validate this change, additional test cases were added to
TestCreateOrUpdate, which exposed a bug in inventoryToConfigMap:
The object statuses were not being parsed. In order to fix this bug,
a new parseDataMap method was added.

To make parsing and formatting easier, the ActuationStrategy,
ActuationStatus, and ReconcileStatus were converted from int enums to
string enums, and the empty string was added as an Unknown constant.
This allows json/yaml formatting and parsing to work round-trip when
the value is the zero value, interpreting the zero value as Unknown
instead of Success. With this change, these enums do not need to
implement MarshalJSON or UnmarshalJSON methods to facilitate formatting
and parsing. Parse methods were also added, to allow enum validation.

With the UnstructuredClient having two ToUnstructuredFunc varients,
the UnstructuredClient no longer needs to know the StatusPolicy, so
StatusPolicy was removed. This clears up ambiguity about what
StatusPolicy meant when the status field was not a subresource. It also
means that only the ToUnstructuredFunc implementer needs to know whether
the status is enabled or disabled. Because of this, the
ConfigMapClientFactory was refactored to use a StatusEnabled bool,
instead of the StatusPolicy.

To facilitate equality comparisons a new ObjectStatusSet alias was added
that wraps []ObjectStatus. It implements the cmp.Equal interface, which
works with testutil.AssertEqual as well, similar to ObjMetadataSet.