perf: Reduce inventory update request size
The UnstructuredClient currently uses the same ToUnstructuredFunc to
generate the spec and the status. This is inefficient when status is
a subresource, because the ToUnstructuredFunc implementation always
needs to return both the spec and the status, which means they both
get sent over the wire to the Kubernetes API. The Kubernetes API will
ignore the spec when using the status endpoint, and ignore the status
when using the normal resource endpoint, so doing all the extra work
to generate, copy, and format the extra fields is inefficient on the
client side. In some edge cases this can also cause errors on the
server side, when the spec and status are larger than the maximum
API request size limit but smaller than the etcd object storage size
limit.

To work around this issue, a new toStatus ToUnstructuredFunc can be
injected into the UnstructuredClient. When non-nil, this new func will
be used before calling UpdateStatus.

This has the added benefit that it makes it possible to use the
UnstructuredClient on custom inventory resources that have a status
field but not a status subresource.

To validate this change, additional test cases were added to
TestCreateOrUpdate, which exposed a bug in inventoryToConfigMap:
The object statuses were not being parsed. In order to fix this bug,
a new parseDataMap method was added.

With the UnstructuredClient having two ToUnstructuredFunc varients,
the UnstructuredClient no longer needs to know the StatusPolicy, so
StatusPolicy was removed. This clears up ambiguity about what
StatusPolicy meant when the status field was not a subresource. It also
means that only the ToUnstructuredFunc implementer needs to know whether
the status is enabled or disabled. Because of this, the
ConfigMapClientFactory was refactored to use a StatusEnabled bool,
instead of the StatusPolicy.